requirements.txt
fastapi
uvicorn
numpy
tensorflow-cpu
keras